fix: remove noise objects from scan tests
