+++ components/log-ingestor/tests/test_ingestion_job.rs
@@ -30,11 +30,7 @@ use log_ingestor::{
     },
 };
 use non_empty_string::NonEmptyString;
-use test_utils::{
-    get_testing_prefix_as_non_empty_string,
-    upload_noise_objects,
-    upload_test_objects,
-};
+use test_utils::{get_testing_prefix_as_non_empty_string, upload_test_objects};
 use tokio::sync::Mutex;
 use uuid::Uuid;
 
@@ -121,6 +117,29 @@ impl S3ScannerState for S3ScannerTestState {
         self.shared_ingested_buffer.lock().await.extend(objects);
         Ok(())
     }
+}
+
+/// Uploads noise S3 objects that do not match any testing prefix.
+///
+/// The keys are formatted as `{uuid}.log`, where `uuid` is a randomly generated v4 UUID.
+///
+/// # Errors
+///
+/// * Forwards [`test_utils::create_s3_objects`]'s return values on failure.
+pub async fn upload_noise_objects(
+    s3_client: aws_sdk_s3::Client,
+    bucket: NonEmptyString,
+    num_objects_to_create: usize,
+) -> Result<()> {
+    let objects_to_create: Vec<_> = (0..num_objects_to_create)
+        .map(|_| ObjectMetadata {
+            bucket: bucket.clone(),
+            key: NonEmptyString::from_string(format!("{}.log", Uuid::new_v4())),
+            size: 16,
+        })
+        .collect();
+
+    test_utils::create_s3_objects(s3_client, objects_to_create).await
 }
 
 /// Waits until the shared buffer has at least `min_num_objects`.
