+++ components/log-ingestor/tests/test_scan.rs
@@ -15,18 +15,13 @@ use clp_rust_utils::{
 };
 use log_ingestor::ingestion_job::scan_prefix;
 use non_empty_string::NonEmptyString;
-use test_utils::{
-    create_s3_objects,
-    get_testing_prefix_as_non_empty_string,
-    upload_noise_objects,
-    upload_test_objects,
-};
+use test_utils::{create_s3_objects, get_testing_prefix_as_non_empty_string, upload_test_objects};
 use tokio::sync::Mutex;
 use uuid::Uuid;
 
+// ListObjectsV2 returns at most 1,000 objects from a bucket
 const NUM_MULTI_PAGE_OBJECTS: usize = 1_205;
 const NUM_FILTERED_OUT_DIRECTORY_MARKERS: usize = 25;
-const NUM_NOISE_OBJECTS: usize = 100;
 
 async fn create_test_s3_client(aws_config: &AwsConfig) -> aws_sdk_s3::Client {
     let aws_auth = AwsAuthentication::Credentials {
@@ -49,8 +44,6 @@ async fn create_test_s3_client(aws_config: &AwsConfig) -> aws_sdk_s3::Client {
 #[ignore = "Requires LocalStack or AWS environment"]
 async fn test_scan_prefix_across_multiple_pages() -> Result<()> {
     let aws_config = AwsConfig::from_env()?;
-    // Prevent unused fields linting.
-    let _ = (&aws_config.account_id, &aws_config.queue_name);
     let s3_client = create_test_s3_client(&aws_config).await;
 
     let job_id = Uuid::new_v4().as_u64_pair().0;
@@ -99,8 +92,6 @@ async fn test_scan_prefix_across_multiple_pages() -> Result<()> {
 #[ignore = "Requires LocalStack or AWS environment"]
 async fn test_scan_prefix_multi_page_filtering() -> Result<()> {
     let aws_config = AwsConfig::from_env()?;
-    // Prevent unused fields linting.
-    let _ = (&aws_config.account_id, &aws_config.queue_name);
     let s3_client = create_test_s3_client(&aws_config).await;
 
     let job_id = Uuid::new_v4().as_u64_pair().0;
@@ -113,21 +104,15 @@ async fn test_scan_prefix_multi_page_filtering() -> Result<()> {
         NUM_MULTI_PAGE_OBJECTS,
     )
     .await?;
-    upload_noise_objects(
-        s3_client.clone(),
-        aws_config.bucket_name.clone(),
-        NUM_NOISE_OBJECTS,
-    )
-    .await?;
 
-    let directory_markers: Vec<_> = (0..NUM_FILTERED_OUT_DIRECTORY_MARKERS)
+    let filtered_out_objects = (0..NUM_FILTERED_OUT_DIRECTORY_MARKERS)
         .map(|idx| ObjectMetadata {
             bucket: aws_config.bucket_name.clone(),
             key: NonEmptyString::from_string(format!("{prefix}/dir-{idx:05}/")),
             size: 16,
         })
         .collect();
-    create_s3_objects(s3_client.clone(), directory_markers).await?;
+    create_s3_objects(s3_client.clone(), filtered_out_objects).await?;
 
     let page_count = Arc::new(AtomicUsize::new(0));
     let scanned_objects = Arc::new(Mutex::new(Vec::<ObjectMetadata>::new()));
